pyproject.toml
[project]
name = "cognee-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cognee>=0.2.0",
    "fastembed>=0.7.1",
    "python-dotenv>=1.1.1",
    "transformers>=4.53.1",
]
